+++ main.go
@@ -6,81 +6,10 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
 
-	"github.com/bep/debounce"
-	corev1 "k8s.io/api/core/v1"
-	v1 "k8s.io/api/networking/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/cache"
 )
-
-type update struct {
-	ingresses []*v1.Ingress
-}
-
-type updateConfigMap struct {
-	configMaps []*corev1.ConfigMap
-}
-
-func listen(ctx context.Context, client kubernetes.Interface, handleUpdate func(*update), handleConfigMapUpdate func(*updateConfigMap)) {
-	factory := informers.NewSharedInformerFactory(client, time.Minute)
-	ingressLister := factory.Networking().V1().Ingresses().Lister()
-	configMapLister := factory.Core().V1().ConfigMaps().Lister()
-
-	onChange := func() {
-		ingresses, err := ingressLister.List(labels.Everything())
-		if err != nil {
-			log.Println("failed to list ingresses: ", err)
-			return
-		}
-		log.Printf("onChange ingress items to review=%d", len(ingresses))
-		handleUpdate(&update{ingresses})
-	}
-
-	onConfigMapChange := func() {
-		configMaps, err := configMapLister.List(labels.Everything())
-		if err != nil {
-			log.Println("failed to list config maps: ", err)
-			return
-		}
-		log.Printf("onChange configmap")
-		handleConfigMapUpdate(&updateConfigMap{configMaps})
-	}
-
-	debounced := debounce.New(time.Second)
-	eventHandler := cache.ResourceEventHandlerFuncs{
-		AddFunc:    func(any) { debounced(onChange) },
-		UpdateFunc: func(any, any) { debounced(onChange) },
-		DeleteFunc: func(any) { debounced(onChange) },
-	}
-
-	eventHandlerConfig := cache.ResourceEventHandlerFuncs{
-		AddFunc:    func(any) { debounced(onConfigMapChange) },
-		UpdateFunc: func(any, any) { debounced(onConfigMapChange) },
-		DeleteFunc: func(any) { debounced(onConfigMapChange) },
-	}
-
-	go func() {
-		i := factory.Networking().V1().Ingresses().Informer()
-		i.AddEventHandler(eventHandler)
-		i.Run(ctx.Done())
-	}()
-	go func() {
-		i := factory.Core().V1().Services().Informer()
-		i.AddEventHandler(eventHandler)
-		i.Run(ctx.Done())
-	}()
-	go func() {
-		i := factory.Core().V1().ConfigMaps().Informer()
-		i.AddEventHandler(eventHandlerConfig)
-		i.Run(ctx.Done())
-	}()
-	<-ctx.Done()
-}
 
 func main() {
 	config, err := rest.InClusterConfig()
@@ -97,7 +26,7 @@ func main() {
 		log.Fatal("TIS: missing TS_AUTHKEY")
 	}
 
-	c := NewHttpController(tsAuthKey)
+	cHttp := NewHttpController(tsAuthKey)
 	cTcp := NewTcpController(tsAuthKey)
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -105,11 +34,13 @@ func main() {
 	signal.Notify(s, syscall.SIGINT, syscall.SIGTERM)
 	go func() {
 		<-s
-		c.shutdown()
+		cHttp.shutdown()
 		cTcp.shutdown()
 		log.Println("shutting down")
 		cancel()
 		os.Exit(0)
 	}()
-	listen(ctx, client, c.update, cTcp.update)
+
+	go cHttp.listen(ctx, client)
+	go cTcp.listen(ctx, client)
 }
